Create a new router for expenses
